flask_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.3.0
folium==0.14.0
pymysql==1.1.0
Werkzeug==2.3.7
